Adding linux support. Had to make a workaround in log.c since linux vfprintf segfaults... I'm fairly sure it's vfprintf at fault, since the file pointer is evidently perfectly valid and the exact same fmt + arguments are used successfully in a call to vfprintf targeting stdout in the line over.

+++ Src/config.h
@@ -72,7 +72,7 @@ typedef int8_t		SALTOFS_T;
 
 #define N_STRIDE 7
 typedef DQWORD_T	WS_T;
-typedef uint64_t	ALU_T;
+typedef uint32_t	ALU_T;
 typedef int32_t		SALTOFS_T;
 
 #define REQUIRED_CAPS	(CPUID_CMOV | CPUID_SSE | CPUID_SSE2)
